Refactored some Card compare

+++ GameCenterTest/CardServiceImpl.m
@@ -14,7 +14,7 @@
 }
 - (void)newCardWithCompletion:(void (^)(Card *))completion {
     [[LinkedInService singleton] getLinkedInPerson:^(LinkedInPerson *person) {
-        Card *card = [Card cardWithName:[NSString stringWithFormat:@"%@ %@", person.firstName, person.lastName] headline:person.headline imageUrl:person.pictureURL.absoluteString connections:person.connections.intValue monthOfEmployment:person.monthOfEmployment.intValue];
+        Card *card = [Card cardWithName:[NSString stringWithFormat:@"%@ %@", person.firstName, person.lastName] headline:person.headline imageUrl:person.pictureURL.absoluteString connections:person.connections monthOfEmployment:person.monthOfEmployment];
         completion(card);
     }                                   andFailure:^(NSString *error) {
         NSLog(@"Could not get linked in connectio %@", error);
